fix error for default hdd|netboot images
commit f811656150ff5f78e55b21702688f082330f78bd enabled the grub-efi
bootloader by default for amd64|i386 architectures, but failed to
recognise the this bootloader is not supported for hdd|netboot images.

this meants that if a user tried to build such an image without explicitly
specifying the bootloader, excluding grub-efi, their build would fail
with an error in the binary_grub-efi stage.

this fixes the problem by only enabling grub-efi by default on supported
image builds.